# apps/api-gateway/src/app/users/documentation/create-user-error-409.serializer.ts
import { ApiProperty } from '@nestjs/swagger';

export class CreateUserError409Serializer {
  @ApiProperty({
    description: 'HTTP code',
    example: 409,
  })
  statusCode: number;

  @ApiProperty({
    description: 'HTTP error message',
    example: 'Conflict',
  })
  message: string;
}